Add King Waypoint
The King is an equally important area to have as a waypoint and marked on the map. Unfortunately, Skyblock does not have a custom location for it like other areas in the Crystal Hollows.

To counteract this, besides setting it manually, I have added a check for messaging starting with "[NPC] King Yolkar" to set the location (and a config to toggle this behavior, if you only want to add King Yolkar manually).