Improve performance of SlayerFeatures.onRenderLivingPre
Avoids instanceof, cast, hasCustomName, getDisplayName, getUnformattedText calls if the feature was not enabled in the first place.

Not sure if any neglible performance gains (there is some because calling hasCustomName or getting displayName calls DataWatcher which has ReentrantReadWriteLock), but it would be bad practice to call methods and assign variables just to do a late if check and exit the method. Moving the config check up after inSkyblock seems like a better practice here.

Signed-off-by: Mustafa ÖNCEL <mustafaoncel92@gmail.com>