perf: Improve performance of SmartFontRenderer.drawString (#543)
* Fix string width

Fixes string width in SmartFontRenderer#drawString.

The Regex is supposed to clean out the control codes in the format of &<code>, to get the raw string that's actually going to be rendered and use that for getStringWidth calculation.

However, the Regex was wrong and only matches the supposedly RGB codes like &[255,  47, 50], I don't understand what is the point in that so its likely just wrong Regex. It was also appearing in Spark reports since its used in rendering code, if you had uncapped FPS limit (no VSync), the .toRegex() in Kotlin creates a new Regex every time drawString is called, which is equivalent to Pattern.compile in Java, which is not a free operation in terms of performance.

This commit just replaces the regex entirely with String#stripControlCodes, which is a Kotlin extension method provided by Utils.kt in Skytils and is used everywhere else in Skytils for removing control codes.

It should both fix the string width being larger than normal due to control codes and improve performance at the same time.

Signed-off-by: Mustafa ÖNCEL <mustafaoncel92@gmail.com>

* Let font render handle the color and formatting chars

FontRenderer already seems to give 0 width for color and most formatting chars, except 1, which is bold, which actually is correct behaviour to add more width if the string is bold.

So just let the font renderer handle this.

Signed-off-by: Mustafa ÖNCEL <mustafaoncel92@gmail.com>

---------

Signed-off-by: Mustafa ÖNCEL <mustafaoncel92@gmail.com>