fix: Ice Fill Solver rendering in boss if unfinished
"Ice Fill" !in DungeonListener.incompletePuzzles hides it correctly once the puzzle is finished, but if the puzzle was not finished, when you go to the boss it will still be rendered.
Adds a new condition to the existing if to ensure bossEntryTime is -1L, meaning boss has not been entered yet, to render the solver.

Also moves config check before inDungeons check (all other checks) as its a simple boolean check and the field only changes if the user changes the option whereas other conditions are more dynamic and might be more expensive.

Signed-off-by: Mustafa ÖNCEL <mustafaoncel92@gmail.com>