perf: Improve condition ordering (#549)
Entity.health in kotlin is a property access and is translated to Entity.getHealth in runtime.
The Entity.getHealth method calls DataWatcher, which uses ReentrantWriteLock, hence the call is not free and the health has to be queried.
Entity.isDead is a real field, thus a simple field read.

It's better to only query the health if isDead is false.

As per the isInvisible - the same thing, it's actually a method call when compiled, and it calls DataWatcher to query the invisibility status just like getHealth.
Instead of going through DataWatcher and ReentrantWriteLock we  are checking the other conditions first, which are simple long comparisions and an instanceof check.
This way we are never calling Entity#isInvisible for non-armorstand entities and when Phase 1 & Boss are cleared (clearTime == -1L).

Signed-off-by: Mustafa ÖNCEL <mustafaoncel92@gmail.com>