perf: Use MRV instead of Inject for modifyTerminalItems
Not sure if this has any other side effects from mixins into the same method, which is why Inject was used in the first place.

However, it stops getStackInSlot from being called twice in the event we return false. Some mods have Mixins into that method that have longer runtimes.

A revert would probably be moving the original.isItemEnchanted check to after the hashset contains.